requirements.txt
aiogram==3.15.0
aiohttp>=3.9.0,<3.11
openai==1.58.1
python-dotenv==1.0.1
qdrant-client==1.12.1
pypdf==5.1.0
python-docx==1.1.2
